requirements.txt
numpy
pandas
matplotlib
pyyaml
scipy
seaborn
